LPS-84119 Move PropertyValueComparator, no logic change
