+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/PropertiesSourceFormatterFileCheck.java
@@ -16,12 +16,11 @@ package com.liferay.source.formatter.checks;
 
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.kernel.util.NaturalOrderStringComparator;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.tools.ToolsUtil;
+import com.liferay.source.formatter.checks.comparator.PropertyValueComparator;
 import com.liferay.source.formatter.checks.util.SourceUtil;
 
 import java.io.File;
@@ -316,52 +315,9 @@ public class PropertiesSourceFormatterFileCheck extends BaseFileCheck {
 		return content;
 	}
 
-	private static final char[][] _REVERSE_ORDER_CHARACTERS = {
-		{CharPool.COLON, CharPool.PERIOD}, {CharPool.DASH, CharPool.SLASH}
-	};
-
 	private static final Pattern _checkPropertyPattern = Pattern.compile(
 		"\n\\s*#?(checkstyle|source\\.check)\\.(.*\\.check)\\.");
 
 	private Boolean _hasPrivateAppsDir;
-
-	private class PropertyValueComparator extends NaturalOrderStringComparator {
-
-		@Override
-		public int compare(String s1, String s2) {
-			int value = super.compare(s1, s2);
-
-			if (s1.startsWith(s2) || s2.startsWith(s1)) {
-				return value;
-			}
-
-			int x = StringUtil.startsWithWeight(s1, s2);
-
-			char c1 = s1.charAt(x);
-			char c2 = s2.charAt(x);
-
-			for (char[] array : _REVERSE_ORDER_CHARACTERS) {
-				if (ArrayUtil.contains(array, c1) &&
-					ArrayUtil.contains(array, c2)) {
-
-					return -value;
-				}
-			}
-
-			if ((x > 0) && (s1.charAt(x - 1) == CharPool.PERIOD)) {
-				if (Character.isUpperCase(c1) && Character.isLowerCase(c2)) {
-					return -1;
-				}
-				else if (Character.isLowerCase(c1) &&
-						 Character.isUpperCase(c2)) {
-
-					return 1;
-				}
-			}
-
-			return value;
-		}
-
-	}
 
 }
